Split consensus from config.
